fixup! libnm-core: use jansson to compare and check team configurations
We expect, that in many cases the strings for
_nm_utils_team_config_equal() are equal. E.g., when user updates
a connection only changing NMSettingIP4Config, but not touching the
team config. It makes sense to check for that first by doing plain
string comparison. In that case, we don't need any special Json
comparison.

Also, adjust the !WITH_JANSSON version, to return the same result
as WITH_JANSSON:
Before:
  WITH:    _nm_utils_team_config_equal (NULL, "", TRUE) => FALSE
  WITHOUT: _nm_utils_team_config_equal (NULL, "", TRUE) => TRUE