sleep-monitor: add functions for delaying the suspension
To allow the execution of asynchronous actions before the system is
suspended, add a mechanism for delaying the drop of inhibitor lock.
Clients can supend the suspension by calling keep_inhibitor() in their
handler for SLEEPING signal and use release_inhibitor() later when
they are done.

We take a "delay" type inhibitor lock, which means that the system
will proceed anyway after a certain amount of time. If some action
were not completed, callers must avoid using release_inhibitor() after
the wake-up.