libnm/vpn: add nm_vpn_plugin_editor_call*() functions
Add function to make a request to a NMVpnPluginEditor generically (by name).
This is useful as it allows adding new functionality to the plugins and use
them without requiring an update of libnm.

IMO, this works around a shortcoming of glib as glib doesn't support
calling methods on an object by name. The closest to that are signals,
but they have the notion of an object notifying others, instead of
invoking a method on the object (and they cannot natively return errors
or multiple values).
Also, you cannot invoke a signal on an object that doesn't implement it
(you'd have to check whether the signal is valid on the object before).