secrets: call CancelGetSecrets after a timeout in secret request
When a call to GetSecrets fails with timeout, we have to notify the
agent that we're not interested in the secrets anymore, so that any
request involving user interaction (as a dialog window) can be
terminated. This was already the normal behavior before commit
92dda6472cb8 ("secret-agent: rework handling of asynchronous request
and cancelling").

Since nm_secret_agent_cancel_secrets() can be now invoked in the
callback of nm_secret_agent_get_secrets(), remove the request from the
hash only after executing the callback, and don't assert that
g_hash_table_remove() succeeds.

Fixes: 92dda6472cb881fc9965a014f16ffabb370b1e55

https://bugzilla.gnome.org/show_bug.cgi?id=767321